initial commit of full-rewrite
