+++ .kitchen/logs/kitchen.log
@@ -1,0 +1,29 @@
+I, [2015-07-16T19:18:12.698092 #32598]  INFO -- Kitchen: -----> Starting Kitchen (v1.4.0)
+E, [2015-07-16T19:18:12.698504 #32598] ERROR -- Kitchen: ------Exception-------
+E, [2015-07-16T19:18:12.698543 #32598] ERROR -- Kitchen: Class: Kitchen::UserError
+E, [2015-07-16T19:18:12.698564 #32598] ERROR -- Kitchen: Message: Kitchen YAML file /Users/ianhenry/Development/chef-slack/.kitchen.yml does not exist.
+E, [2015-07-16T19:18:12.698584 #32598] ERROR -- Kitchen: ------Backtrace-------
+E, [2015-07-16T19:18:12.698626 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/loader/yaml.rb:74:in `read'
+E, [2015-07-16T19:18:12.698640 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/config.rb:145:in `data'
+E, [2015-07-16T19:18:12.698653 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/config.rb:124:in `suites'
+E, [2015-07-16T19:18:12.698665 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/config.rb:175:in `filter_instances'
+E, [2015-07-16T19:18:12.698676 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/config.rb:134:in `build_instances'
+E, [2015-07-16T19:18:12.698687 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/config.rb:110:in `instances'
+E, [2015-07-16T19:18:12.698699 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/command.rb:115:in `filtered_instances'
+E, [2015-07-16T19:18:12.698710 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/command.rb:145:in `parse_subcommand'
+E, [2015-07-16T19:18:12.698742 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/command/action.rb:38:in `block in call'
+E, [2015-07-16T19:18:12.698754 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/lib/ruby/2.1.0/benchmark.rb:279:in `measure'
+E, [2015-07-16T19:18:12.698765 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/command/action.rb:37:in `call'
+E, [2015-07-16T19:18:12.698776 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/cli.rb:56:in `perform'
+E, [2015-07-16T19:18:12.698787 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/cli.rb:171:in `block (2 levels) in <class:CLI>'
+E, [2015-07-16T19:18:12.698798 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/lib/ruby/gems/2.1.0/gems/thor-0.19.1/lib/thor/command.rb:27:in `run'
+E, [2015-07-16T19:18:12.698809 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/lib/ruby/gems/2.1.0/gems/thor-0.19.1/lib/thor/invocation.rb:126:in `invoke_command'
+E, [2015-07-16T19:18:12.698820 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/cli.rb:308:in `invoke_task'
+E, [2015-07-16T19:18:12.698831 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/lib/ruby/gems/2.1.0/gems/thor-0.19.1/lib/thor.rb:359:in `dispatch'
+E, [2015-07-16T19:18:12.698842 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/lib/ruby/gems/2.1.0/gems/thor-0.19.1/lib/thor/base.rb:440:in `start'
+E, [2015-07-16T19:18:12.698853 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/bin/kitchen:13:in `block in <top (required)>'
+E, [2015-07-16T19:18:12.698864 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/lib/kitchen/errors.rb:154:in `with_friendly_errors'
+E, [2015-07-16T19:18:12.698875 #32598] ERROR -- Kitchen: /opt/chefdk/embedded/apps/test-kitchen/bin/kitchen:13:in `<top (required)>'
+E, [2015-07-16T19:18:12.698904 #32598] ERROR -- Kitchen: /usr/bin/kitchen:15:in `load'
+E, [2015-07-16T19:18:12.698915 #32598] ERROR -- Kitchen: /usr/bin/kitchen:15:in `<main>'
+E, [2015-07-16T19:18:12.698925 #32598] ERROR -- Kitchen: ----------------------
