+++ CHANGELOG.md
@@ -1,14 +1,6 @@
 slack CHANGELOG
 ===============
 
-0.1.2
------
-- [James Wickett] - locked to new Slackr gem 0.0.5 and now using the new syntax Slackr.connect
-
-0.1.1
------
-- [James Wickett] - bugfix Issue #1 and updated the LWRP to be slack_say 
-
 0.1.0
 -----
-- [Jason Rohwedder] - Initial release of slack
+- [Ian Henry] - Initial release of slack
